+++ src/FamilyHubs.SharedKernel.Razor/Dashboard/ICell.cs
@@ -13,9 +13,4 @@ public interface ICell
     /// If PartialName is null, this will be rendered as the cell's content. It can contain HTML.
     /// </summary>
     string? ContentAsHtml { get; }
-
-    /// <summary>
-    /// The (space separated) classes to be applied to the cell element.
-    /// </summary>
-    //public string? Classes { get; }
 }